@@ -1,6 +1,6 @@
 {
   "name": "simple-update-notifier",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "Simple update notifier to check for npm updates for cli applications",
   "main": "build/index.js",
   "types": "build/index.d.ts",
@@ -12,11 +12,11 @@
   "author": "alexbrazier",
   "license": "MIT",
   "engines": {
-    "node": ">=8.10.0"
+    "node": ">=10"
   },
   "scripts": {
     "test": "jest src --noStackTrace",
-    "build": "rollup -c rollup.config.js",
+    "build": "rollup -c rollup.config.js --bundleConfigAsCjs",
     "prettier:check": "prettier --check src/**/*.ts",
     "prettier": "prettier --write src/**/*.ts",
     "eslint": "eslint src/**/*.ts",
@@ -25,25 +25,28 @@
     "release": "release-it"
   },
   "dependencies": {
-    "semver": "~7.0.0"
+    "semver": "^7.5.3"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.19.1",
-    "@babel/preset-typescript": "^7.17.12",
-    "@release-it/conventional-changelog": "^5.1.0",
-    "@types/jest": "^29.0.3",
-    "@types/node": "^18.7.18",
-    "@typescript-eslint/eslint-plugin": "^5.37.0",
-    "@typescript-eslint/parser": "^5.37.0",
-    "eslint": "^8.23.1",
-    "eslint-config-prettier": "^8.5.0",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
+    "@release-it/conventional-changelog": "^5.1.1",
+    "@types/jest": "^29.5.2",
+    "@types/node": "^20.3.1",
+    "@typescript-eslint/eslint-plugin": "^5.60.0",
+    "@typescript-eslint/parser": "^5.60.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^29.0.3",
-    "prettier": "^2.7.1",
-    "release-it": "^15.4.2",
-    "rollup": "^2.79.0",
-    "rollup-plugin-ts": "^3.0.2",
-    "typescript": "^4.8.3"
+    "jest": "^29.5.0",
+    "prettier": "^2.8.8",
+    "release-it": "^15.11.0",
+    "rollup": "^3.25.2",
+    "rollup-plugin-ts": "^3.2.0",
+    "typescript": "^5.1.3"
+  },
+  "resolutions": {
+    "semver": "^7.5.3"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
