@@ -1,14 +1,12 @@
 import express from "express";
 import dotenv from "dotenv";
 import cors from "cors";
-// Mongo connection removed
 import { notFound, errorHandler } from "./middleware/errorMiddleware.js";
 
-// Route files
+// Routes
 import userRoutes from "./routes/userRoutes.js";
 import platformRoutes from "./routes/platformRoutes.js";
 import verificationRoutes from "./routes/verificationRoutes.js";
-// Removed legacy Mongo routes
 
 dotenv.config();
 
@@ -22,23 +20,20 @@ const allowedOrigins = [
   'https://localhost:3000',
 ].filter(Boolean);
 
-// Helper to decide if an origin is allowed (supports *.vercel.app previews)
 function isAllowedOrigin(origin) {
-  if (!origin) return true; // non-browser or same-origin
+  if (!origin) return true;
   if (allowedOrigins.includes(origin)) return true;
   try {
     const url = new URL(origin);
     const hostname = url.hostname;
     if (hostname === 'localhost') return true;
-    // Allow any Vercel preview or production subdomain
     if (hostname.endsWith('.vercel.app')) return true;
-    // Optional: allow CSV list via EXTRA_CORS_ORIGINS
     if (process.env.EXTRA_CORS_ORIGINS) {
       const extras = process.env.EXTRA_CORS_ORIGINS.split(',').map(s => s.trim()).filter(Boolean);
       if (extras.includes(origin)) return true;
     }
-  } catch (_) {
-    // ignore parse errors; fall through to deny
+  } catch {
+    // ignore parse errors
   }
   return false;
 }
@@ -52,18 +47,16 @@ app.use(
     credentials: true,
   })
 );
+
 app.use(express.json());
 
-// API Routes
 app.use("/api/users", userRoutes);
 app.use("/api/platforms", platformRoutes);
 app.use("/api/verification", verificationRoutes);
 
-// Healthcheck
 app.get("/", (req, res) => res.send("CodeDesk API is running"));
 
 app.use(notFound);
 app.use(errorHandler);
 
-const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => console.log(`Server running on port ${PORT}`)); 
+export default app;
