@@ -1,9 +1,9 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-  plugins: [react()],
-  server: {
-    port: 3000,
+  test: {
+    globals: true,
+    environment: "node",
+    include: ["tests/**/*.mjs"], // only run .mjs tests
   },
-}); 
+});
